Fix some clang warnings.
* graph/grfill.c (grFillRect): Use `int' for `size'.

* src/ftstring.c (flower_init): Use `unsigned int' for `i'.
Simplify some loops.

* src/ftview.c (parse_cmdline) <'L'>: Use `unsigned char' casts.
(main): Add `FT_LcdFilter' cast.

+++ graph/grfill.c
@@ -231,7 +231,7 @@ grFillRect( grBitmap*   target,
   int              delta;
   unsigned char*   line;
   grFillHLineFunc  hline_func;
-  size_t           size = 0;
+  int              size = 0;
 
   if ( x < 0 )
   {
@@ -271,7 +271,7 @@ grFillRect( grBitmap*   target,
     size += 2;
     hline_func( line, x, width, color );
     for ( line += size * x; --height > 0; line += target->pitch )
-      memcpy( line + target->pitch, line, size * width );
+      memcpy( line + target->pitch, line, (size_t)size * (size_t)width );
     break;
 
   case gr_pixel_mode_gray:
